requirements.txt
selenium-driverless
asyncio
loguru
pyyaml